feat: add put endpoint for BLOBs (#61)
* feat: add put endpoint for BLOBs

* feat: add blob puts client snapshot code

* fix: rm zealous check

* feat: human_size pretty printer

* fix: relpath is now optional on snapshots

This is so that it doesn't spew annoying errors when you try to take a snapshot of some data that is outside workspace.

* feat: first pass of type-dispatched validator

* test: snapshot for humansized

* fix: hacky fix to solve `Payload: !Send` synchronization issue

This fix sets the HTTP server to single-threaded, which is not a solution at all. Waiting on a better fix. See https://github.com/awslabs/aws-sdk-rust/discussions/458#discussioncomment-4022492

* fix: use store config

* feat: add HEAD method for `/blob`

* fix: codedeps in classes

It's still buggy spaghetti but it does the job

* fix test

Co-authored-by: E.W.Ayers <edward.ayers@outlook.com>