fix: result_json too big
fixes #82. The problem was that the size of results_json could be unbounded.
I fixed this by truncating after 100 items in dictionaries or lists.

I think we need a better answer here, so I've made a hacky fix for now.
The problem to solve is how to make a nice visualisation that doesn't use up loads of space by default, but which can be expanded into a full thing if needed.

The ideal solution is to have xyz/web be able to view the pickles directly and only load the first megabyte or whatever and show that.