ci: deployment (#102)
The GitHub OAuth system requires us to provide a redirect_url parameter, which is where the user's browser goes after they successfully completed the login flow on GitHub's side. Untili now, we've had this be `127.0.0.1:PORT` when called from the Python client. From the web application, we've used `127.0.0.1:3000/login`. This has been fine, because both of these have the same host name (`127.0.0.1`). However, in production, they can't have the same hostname, because the Python client is trying to redirect to the miniserver on the user's machine, while the web app needs to redirect to a live web page at hitsave.io. The problem is that the GitHub OAuth system will only accept one base redirect url, not multiple.

To solve this, we are now doing the following, which is actually somewhat simpler. The GitHub registered `redirect_url` (also known as callback URL) will be the live hosted hitsave.io page at `https://hitsave.io/login`. This page can accept additional search params, so we are adding a `client_loopback` param, which will include the loopback URL to the miniserver. In this way, we can use the Remix website to display the pretty login page that we've always wanted, and then use a `fetch` request from the locally rendered component in a `useEffect` hook to ping the miniserver at the URL which was passed in `client_loopback`. This new approach removes the need for the Python client to communicate with the API to pass the `code` parameter through and get the user registered. All of this now happens from the NodeJS hosted web app, reusing the same code which does this for browser-initiated logins.

This commit implements all of this.